Commit Before PyOpenGL Overhaul
Over the past few days I've been experimenting with OpenGL and PyOpenGL, namely with the goal of understanding Vertex Buffer Objects in order to take advantage of the GPU. I have managed to put together a simple rotating cuboid demo using VBOs for the vertices and triangles, but I did so in external sandbox files and "merging" will involve a lot of deletions and overwrites. I want to ensure that I have a record of these changes.

+++ nsvt_config.py
@@ -7,9 +7,9 @@ SHAPE3D_VERTICES_NUMPY_DTYPE = numpy.float32
 # NumPy Datatype for the Edges array
 #  - Values will always be zero or positive, so the type can be unsigned
 #  - uint8 will only support up to 256 vertices in a shape, whereas uint16 allows up to 65536
-SHAPE3D_EDGES_NUMPY_DTYPE = numpy.uint16
+SHAPE3D_EDGES_NUMPY_DTYPE = numpy.uint32
 
 # NumPy Datatype for the Triangles array
 #  - Values will always be zero or positive, so the type can be unsigned
 #  - uint8 will only support up to 256 vertices in a shape, whereas uint16 allows up to 65536
-SHAPE3D_TRIANGLES_NUMPY_DTYPE = numpy.uint16
+SHAPE3D_TRIANGLES_NUMPY_DTYPE = numpy.uint32
